Naive Bayes Classifier Clustering Heatmap
This clustering heatmap shows the results of the Naive Bayes Classifier on the 674 samples I used in the model (I removed 4 profiles that were difficult to pre-process).  Rows are individual samples, with the left annotation strip showing (by different color) the true number being spoken.  The columns represent the possible classifications assigned by the classifier.  The cells of the heatmap show the posterior probability of being classified as the label indicated at the bottom of the columns (intense blue = 100%, white = 0%).  If you assign each sample the label corresponding to the maximum probability, the naive error rate is 18.5%.  10-fold cross-validated error rate is identical.